Version 1.65.0
•	Minor fixes in AboutWindow
	o	Remove old window.
	o	Add copyright text.
•	Move save/load of geometries from SystemStructure to RealGeometry.
•	Encapsulate data in RealGeometry.
•	Move rotation matrix to BaseGeometry
•	Unify movement of geometries on CPU and GPU.
	o	Allow geometry velocity and free motion together on GPU, the same as on CPU
•	AnalysisVolume encapsulation.
	o	Move some functionality from SystemStructure to AnalysisVolume.
	o	Move mesh generation into separate class.
	o	Class for managing sizes of geometries.
•	RealGeometry encapsulation.
	o	Move some functionality from SystemStructure to RealGeometry.
•	Use BaseGeometry in GeometriesEditorTab.
•	Fix and tune TableItemSpinBox.
•	Add missing copyright notices
•	Updated documentation for geometries editor and time points
•	Updates in GeometriesEditorTab.
	o	Show list of geometries with a single tree widget.
	o	Fix proto-bug with setting mesh in RealGeometry.
	o	Add new QtTree widget.
•	Show geometry properties in GeometriesEditor via TreeWidget.
•	Dynamic generator with PBC
	o	Generation of agglomerates with dynamic generator accounts for PBC's and allows generation over PBCs.
	o	Generation volume has to be larger than PBC volume; if equal, will be considered as wall.
	o	Fix possibly missing Verlet update after the generation of objects.
	o	Adjust simulation domain in VerletList if PBC is enabled to include all possible virtual particles.
•	Update ObjectsGeneratorTab
	o	Possibility to set Max Iterations Number.
	o	Possibility to define Generation Rate as: generation rate, objects per step, or total objects.
	o	Slight refactoring of ObjectsGeneratorTab.
•	Check material parameters in MBD for invalid values before starting the simulation.
•	Saving/loading of tangential overlaps for solid bonds
•	Extend units converter
	o	Add ANGLE to units converter.
	o	Fix bug with ANGULAR_VELOCITY.
•	Extended comparison of mdem files

+++ Databases/AgglomerateDatabase/AgglomeratesDatabase.cpp
@@ -25,7 +25,7 @@ void CAgglomeratesDatabase::AddNewAgglomerate(SAgglomerate& _pAgglomerate)
 	m_vAgglomerates.push_back(pAgglom);
 }
 
-SAgglomerate* CAgglomeratesDatabase::GetAgglomerate(unsigned _nIndex)
+SAgglomerate* CAgglomeratesDatabase::GetAgglomerate(size_t _nIndex)
 {
 	if (_nIndex >= m_vAgglomerates.size()) return NULL;
 	return m_vAgglomerates[_nIndex];
